add debug to worker

+++ worker.php
@@ -8,5 +8,8 @@ if (isset($argv[1]) && file_exists(WORKER_CACHE_PATH . '/' . $argv[1])) {
   $data = json_decode($data, true);
   $reply = doLogic($data);
   sendMessageWithRetry($reply);
-  unlink(WORKER_CACHE_PATH . '/' . $argv[1]);
+
+  if (! isset($argv[2]) || $argv[2] !== 'debug') {
+    unlink(WORKER_CACHE_PATH . '/' . $argv[1]);
+  }
 }
